src/requirements.txt
django==3.2.0
djangorestframework==3.15.0
django-cors-headers==4.3.1
requests
selenium==4.12.0
PyYAML==3.11
image==1.5.33
pandas==1.1.5
numpy==1.25.1
opencv-python==4.9.0.80
cached-property==1.5.2
pyperclip==1.9.0
openpyxl==3.1.2
